+++ packages/react-hotkeys-devtools/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @tanstack/react-hotkeys-devtools
 
+## 0.2.0
+
+### Patch Changes
+
+- Updated dependencies [[`8ae6b64`](https://github.com/TanStack/hotkeys/commit/8ae6b64ef10b53186c367f594f81ecdec15071d7)]:
+  - @tanstack/hotkeys-devtools@0.2.0
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/hotkeys-devtools@0.1.3
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/hotkeys-devtools@0.1.2
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/hotkeys-devtools@0.1.1
+
 ## 0.1.0
 
 ### Patch Changes