+++ packages/react-hotkeys/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @tanstack/react-hotkeys
 
+## 0.2.0
+
+### Minor Changes
+
+- feat: upgrade tanstack store version ([#35](https://github.com/TanStack/hotkeys/pull/35))
+
+### Patch Changes
+
+- Updated dependencies [[`8ae6b64`](https://github.com/TanStack/hotkeys/commit/8ae6b64ef10b53186c367f594f81ecdec15071d7)]:
+  - @tanstack/hotkeys@0.2.0
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [[`26a74d8`](https://github.com/TanStack/hotkeys/commit/26a74d8e3279766a5cddbc7e7f146af0557cfbb9)]:
+  - @tanstack/hotkeys@0.1.3
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [[`54f550f`](https://github.com/TanStack/hotkeys/commit/54f550f1d2e47084ea11411d465780ede19fbdfa)]:
+  - @tanstack/hotkeys@0.1.2
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [[`56d50ba`](https://github.com/TanStack/hotkeys/commit/56d50ba4595a7609b13a36376afdc46ba92fab29)]:
+  - @tanstack/hotkeys@0.1.1
+
 ## 0.1.0
 
 ### Minor Changes